Added paths verification

+++ contest.py
@@ -2,6 +2,7 @@ from latexutils import clean_auxiliary_files
 from pdfutils import build_pdf, merge_pdfs
 from boca import boca_pack
 from utils import convert_idx_to_string
+import subprocess
 import sys
 import os
 import argparse
@@ -31,8 +32,8 @@ def build_contest_pdf(problem_folder_l, output_folder):
     problem_pdf_l = []
     tutorial_pdf_l = []
 
-    cls_file = os.path.join(os.path.dirname(__file__), 'arquivos')
-    cls_file = os.path.join(cls_file, 'maratona.cls')
+    cls_file = os.path.join(os.path.join(os.path.dirname(
+        os.path.abspath(__file__)), 'arquivos'), 'maratona.cls')
     shutil.copy(cls_file, output_folder)
 
     for i, folder in enumerate(problem_folder_l):
@@ -53,7 +54,8 @@ def build_contest_pdf(problem_folder_l, output_folder):
         os.remove(f)
     for f in tutorial_pdf_l:
         os.remove(f)
-    os.remove(os.path.join(output_folder, 'maratona.cls'))
+    if output_folder not in problem_folder_l:
+        os.remove(os.path.join(output_folder, 'maratona.cls'))
 
 
 """
@@ -78,17 +80,28 @@ if __name__ == '__main__':
     parser = create_parser()
     args = parser.parse_args()
 
-    for problems in args.problem_path:
-        if (not os.path.exists(problems)):
-            print(problems, "path doesn't exist.")
+    for problem in args.problem_path:
+        if (not os.path.exists(problem)):
+            print(problem, "path doesn't exist.")
             sys.exit(1)
-        if (not os.path.exists(os.path.join(problems, 'statement.md'))):
-            print(problems, "path doesn't have an initialized problem.")
+        if (not os.path.exists(os.path.join(problem, 'statement.md'))):
+            print(problem, "path doesn't have an initialized problem.")
             sys.exit(1)
-        if (not os.path.exists(os.path.join(problems, 'bin'))):
-            print(problems, "path doesn't have a built problem.")
-            sys.exit(1)
-    
+        if (args.mode == 'genpdf'):
+            if (not os.path.exists(os.path.join(problem, 'input'))):
+                print(problem, "path doesn't have an input folder.")
+                sys.exit(1)
+            if (not os.path.exists(os.path.join(problem, 'output'))):
+                print(problem, "path doesn't have an output folder.")
+                sys.exit(1)
+        elif (args.mode == 'build' and not os.path.exists(os.path.join(problem, 'bin'))):
+            command = ['python3', os.path.join(os.path.dirname(os.path.relpath(__file__)), 
+                    'build.py'), 'build', problem]
+            p = subprocess.run(command, stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+            if (p.stderr):
+                print("Error building problem.")
+                sys.exit(1)
+
     os.makedirs(args.contest_folder, exist_ok=True)
 
     if (args.mode == 'build' and args.boca):
